requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
redis[hiredis]==5.0.4
asyncpg==0.29.0
prometheus-client==0.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
locust==2.27.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
pytest-asyncio==0.23.6
